refactor(commit): move git related code into its own module
